+feat: Refactor authentication service to use SysRoleMapper and update role names in schema.sql

+++ wenxige-bg-service/src/main/java/com/wenxige/bg/entity/SysUser.java
@@ -36,7 +36,7 @@ public class SysUser implements UserDetails {
 
     private String email;
 
-    private Integer sysRoleId;
+    private Long sysRoleId;
 
     private String createUser;
